refactor(markdown): relocate parser to lib/parsers/markdown; mirror tests under tests/parsers/markdown
- Move types/inline/blocks/references/frontmatter/index from
  tests/examples/markdown to lib/parsers/markdown so the parser is
  shipped as part of the library build (matches the existing
  lib/parsers/within.ts precedent).
- Move the per-file test files into tests/parsers/markdown so the test
  tree mirrors the source tree, consistent with the rest of the suite.
- Convert @/lib/* imports in the source files to relative ../../* with
  .js suffixes so the emitted dist/ resolves correctly under Node.
- Drop the legacy monolithic tests/examples/markdown.test.ts whose 18
  tests duplicate coverage already in the split per-file tests (this
  also clears the pre-existing vitest.globals tsc error).
- Fix the README example link to point at the new lib/ location.

Amp-Thread-ID: https://ampcode.com/threads/T-019e70f5-50bf-70bc-ba46-babf0645a026
